Refactor screens that comprise the password reset process
