+++ WasHere/src/navigation/AuthNavigator.js
@@ -3,9 +3,9 @@ import { createStackNavigator } from '@react-navigation/stack';
 
 import Welcome from '../screens/Welcome';
 import Login from '../screens/Login';
-import ResetForm from '../screens/ForgotPassword/ResetForm';
-import CodeForm from '../screens/ForgotPassword/CodeForm';
-import ChangePasswordForm from '../screens/ForgotPassword/ChangePasswordForm';
+import ForgottenPassword from '../screens/ForgottenPassword';
+import ResetCode from '../screens/ResetCode';
+import ChangePassword from '../screens/ChangePassword';
 
 const Stack = createStackNavigator();
 
@@ -13,9 +13,9 @@ const AuthNavigator = () => (
   <Stack.Navigator screenOptions={{ animationEnabled: false }}>
     <Stack.Screen name="Welcome" component={Welcome} options={{ headerShown: false }} />
     <Stack.Screen name="Login" component={Login} options={{ headerShown: false }} />
-    <Stack.Screen name="ResetPasswordForm" component={ResetForm} options={{ headerShown: false }} />
-    <Stack.Screen name="CodeForm" component={CodeForm} options={{ title: 'Verify reset code' }} />
-    <Stack.Screen name="ResetChangePasswordForm" component={ChangePasswordForm} options={{ headerShown: false }} />
+    <Stack.Screen name="ResetPasswordForm" component={ForgottenPassword} options={{ headerShown: false }} />
+    <Stack.Screen name="CodeForm" component={ResetCode} options={{ title: 'Verify reset code' }} />
+    <Stack.Screen name="ResetChangePasswordForm" component={ChangePassword} options={{ headerShown: false }} />
   </Stack.Navigator>
 );
